requirements.txt
python-dotenv
langchain
openai
langchain_openai
pandas
streamlit
streamlit-modal
Flask
scikit-learn
xgboost
gunicorn
openpyxl
flask-cors